Refactor: split out flavor assigment logic
