+++ pkg/scheduler/scheduler.go
@@ -20,7 +20,6 @@ import (
 	"context"
 	"fmt"
 	"sort"
-	"strings"
 	"time"
 
 	"github.com/go-logr/logr"
@@ -32,8 +31,6 @@ import (
 	"k8s.io/apimachinery/pkg/util/sets"
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/client-go/tools/record"
-	corev1helpers "k8s.io/component-helpers/scheduling/corev1"
-	"k8s.io/component-helpers/scheduling/corev1/nodeaffinity"
 	"k8s.io/klog/v2"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -43,6 +40,7 @@ import (
 	"sigs.k8s.io/kueue/pkg/constants"
 	"sigs.k8s.io/kueue/pkg/metrics"
 	"sigs.k8s.io/kueue/pkg/queue"
+	"sigs.k8s.io/kueue/pkg/scheduler/flavorassigner"
 	"sigs.k8s.io/kueue/pkg/util/api"
 	"sigs.k8s.io/kueue/pkg/util/routine"
 	"sigs.k8s.io/kueue/pkg/workload"
@@ -118,7 +116,7 @@ func (s *Scheduler) schedule(ctx context.Context) {
 			continue
 		}
 		c := snapshot.ClusterQueues[e.ClusterQueue]
-		if len(e.borrows) > 0 && c.Cohort != nil && usedCohorts.Has(c.Cohort.Name) {
+		if e.assignment.Borrows() && c.Cohort != nil && usedCohorts.Has(c.Cohort.Name) {
 			e.status = skipped
 			e.inadmissibleMsg = "cohort used in this cycle"
 			continue
@@ -169,9 +167,7 @@ type entry struct {
 	// workload.Info holds the workload from the API as well as resource usage
 	// and flavors assigned.
 	workload.Info
-	// borrows is the resources that the workload would need to borrow from the
-	// cohort if it was scheduled in the clusterQueue.
-	borrows         cache.ResourceQuantities
+	assignment      flavorassigner.Assignment
 	status          entryStatus
 	inadmissibleMsg string
 	requeueReason   queue.RequeueReason
@@ -196,116 +192,15 @@ func (s *Scheduler) nominate(ctx context.Context, workloads []workload.Info, sna
 		} else if !cq.NamespaceSelector.Matches(labels.Set(ns.Labels)) {
 			e.inadmissibleMsg = "Workload namespace doesn't match ClusterQueue selector"
 			e.requeueReason = queue.RequeueReasonNamespaceMismatch
-		} else if status := e.assignFlavors(log, snap.ResourceFlavors, cq); !status.IsSuccess() {
+		} else if assignment, status := flavorassigner.AssignFlavors(log, &e.Info, snap.ResourceFlavors, cq); !status.IsSuccess() {
 			e.inadmissibleMsg = api.TruncateEventMessage(status.Message())
 		} else {
+			e.assignment = *assignment
 			e.status = nominated
 		}
 		entries = append(entries, e)
 	}
 	return entries
-}
-
-type admissionStatus struct {
-	podSet  string
-	reasons []string
-	err     error
-}
-
-// Message returns a concatenated message on reasons of the admissionStatus.
-func (s *admissionStatus) Message() string {
-	if s.IsSuccess() {
-		return ""
-	}
-	if s.IsError() {
-		return fmt.Sprintf("Couldn't assign flavors for podSet %s: %v", s.podSet, s.err)
-	}
-	sort.Strings(s.reasons)
-	msg := strings.Join(s.reasons, "; ")
-	return fmt.Sprintf("Workload's %q podSet didn't fit: %s", s.podSet, msg)
-}
-
-// AppendReason appends given reasons to the admissionStatus.
-func (s *admissionStatus) AppendReason(reasons ...string) {
-	s.reasons = append(s.reasons, reasons...)
-}
-
-// IsSuccess returns true if "admissionStatus" is nil.
-func (s *admissionStatus) IsSuccess() bool {
-	return s == nil
-}
-
-// IsError returns true if "admissionStatus" has error.
-func (s *admissionStatus) IsError() bool {
-	return s != nil && s.err != nil
-}
-
-// asStatus wraps an error in a admissionStatus.
-func asStatus(err error) *admissionStatus {
-	return &admissionStatus{
-		err: err,
-	}
-}
-
-// assignFlavors calculates the flavors that should be assigned to this entry
-// if admitted by this clusterQueue, including details of how much it needs to
-// borrow from the cohort.
-// It returns admissionStatus indicating whether the entry fits. If it doesn't fit,
-// the entry is unmodified.
-func (e *entry) assignFlavors(log logr.Logger, resourceFlavors map[string]*kueue.ResourceFlavor, cq *cache.ClusterQueue) *admissionStatus {
-	flavoredRequests := make([]workload.PodSetResources, 0, len(e.TotalRequests))
-	wUsed := make(cache.ResourceQuantities)
-	wBorrows := make(cache.ResourceQuantities)
-	for i, podSet := range e.TotalRequests {
-		assignedFlavors := make(map[corev1.ResourceName]string, len(podSet.Requests))
-		for resName := range podSet.Requests {
-			if assignedFlavors[resName] != "" {
-				// This resource got assigned the same flavor as a codependent resource.
-				// No need to compute again.
-				continue
-			}
-			if _, ok := cq.RequestableResources[resName]; !ok {
-				return &admissionStatus{
-					reasons: []string{fmt.Sprintf("resource %s unavailable in ClusterQueue", resName)},
-				}
-			}
-			codepResources := cq.RequestableResources[resName].CodependentResources
-			if codepResources.Len() == 0 {
-				codepResources = sets.NewString(string(resName))
-			}
-			codepReq := filterRequestedResources(podSet.Requests, codepResources)
-			rFlavor, borrows, status := findFlavorForCodepResources(log, codepReq, wUsed, resourceFlavors, cq, &e.Obj.Spec.PodSets[i].Spec)
-			if !status.IsSuccess() {
-				status.podSet = e.Obj.Spec.PodSets[i].Name
-				return status
-			}
-			for codepRes, codepVal := range codepReq {
-				if b := borrows[codepRes]; b > 0 {
-					if wBorrows[codepRes] == nil {
-						wBorrows[codepRes] = make(map[string]int64)
-					}
-					// Don't accumulate borrowing. The returned `borrow` already considers
-					// usage from previous pod sets.
-					wBorrows[codepRes][rFlavor] = b
-				}
-				if wUsed[codepRes] == nil {
-					wUsed[codepRes] = make(map[string]int64)
-				}
-				wUsed[codepRes][rFlavor] += codepVal
-				assignedFlavors[codepRes] = rFlavor
-			}
-		}
-		flavoredRequests = append(flavoredRequests, workload.PodSetResources{
-			Name:     podSet.Name,
-			Requests: podSet.Requests,
-			Flavors:  assignedFlavors,
-		})
-	}
-	e.TotalRequests = flavoredRequests
-	if len(wBorrows) > 0 {
-		e.borrows = wBorrows
-	}
-	return nil
 }
 
 // admit sets the admitting clusterQueue and flavors into the workload of
@@ -316,13 +211,7 @@ func (s *Scheduler) admit(ctx context.Context, e *entry) error {
 	newWorkload := e.Obj.DeepCopy()
 	admission := &kueue.Admission{
 		ClusterQueue:  kueue.ClusterQueueReference(e.ClusterQueue),
-		PodSetFlavors: make([]kueue.PodSetFlavors, len(e.TotalRequests)),
-	}
-	for i := range e.TotalRequests {
-		admission.PodSetFlavors[i] = kueue.PodSetFlavors{
-			Name:    e.Obj.Spec.PodSets[i].Name,
-			Flavors: e.TotalRequests[i].Flavors,
-		}
+		PodSetFlavors: e.assignment.ToAPI(),
 	}
 	newWorkload.Spec.Admission = admission
 	if err := s.cache.AssumeWorkload(newWorkload); err != nil {
@@ -383,152 +272,6 @@ func workloadAdmissionFrom(w *kueue.Workload) *kueue.Workload {
 	return wlCopy
 }
 
-// findFlavorForCodepResources returns a flavor which can satisfy the resource request,
-// given that wUsed is the usage of flavors by previous podsets.
-// If it finds a flavor, also returns any borrowing required.
-func findFlavorForCodepResources(
-	log logr.Logger,
-	requests workload.Requests,
-	wUsed cache.ResourceQuantities,
-	resourceFlavors map[string]*kueue.ResourceFlavor,
-	cq *cache.ClusterQueue,
-	spec *corev1.PodSpec) (string, map[corev1.ResourceName]int64, *admissionStatus) {
-	var status admissionStatus
-
-	// Keep any resource name as an anchor to gather flavors for.
-	var rName corev1.ResourceName
-	for rName = range requests {
-		break
-	}
-
-	// We will only check against the flavors' labels for the resource.
-	// Since all the resources share the same flavors, they use the same selector.
-	selector := flavorSelector(spec, cq.LabelKeys[rName])
-	for i, flvLimit := range cq.RequestableResources[rName].Flavors {
-		flavor, exist := resourceFlavors[flvLimit.Name]
-		if !exist {
-			log.Error(nil, "Flavor not found", "Flavor", flvLimit.Name)
-			status.AppendReason(fmt.Sprintf("flavor %s not found", flvLimit.Name))
-			continue
-		}
-		taint, untolerated := corev1helpers.FindMatchingUntoleratedTaint(flavor.Taints, spec.Tolerations, func(t *corev1.Taint) bool {
-			return t.Effect == corev1.TaintEffectNoSchedule || t.Effect == corev1.TaintEffectNoExecute
-		})
-		if untolerated {
-			status.AppendReason(fmt.Sprintf("untolerated taint %s in flavor %s", taint, flvLimit.Name))
-			continue
-		}
-		if match, err := selector.Match(&corev1.Node{ObjectMeta: metav1.ObjectMeta{Labels: flavor.NodeSelector}}); !match || err != nil {
-			if err != nil {
-				return "", nil, asStatus(fmt.Errorf("matching affinity flavor %s: %w", flvLimit.Name, err))
-			}
-			status.AppendReason(fmt.Sprintf("flavor %s doesn't match with node affinity", flvLimit.Name))
-			continue
-		}
-
-		fitsAll := true
-		borrows := make(map[corev1.ResourceName]int64, len(requests))
-		for name, val := range requests {
-			codepFlvLimit := cq.RequestableResources[name].Flavors[i]
-			// Check considering the flavor usage by previous pod sets.
-			borrow, s := fitsFlavorLimits(name, val+wUsed[name][flavor.Name], cq, &codepFlvLimit)
-			if s.IsError() {
-				return "", nil, s
-			}
-			if !s.IsSuccess() {
-				fitsAll = false
-				status.AppendReason(s.reasons...)
-				break
-			}
-			borrows[name] = borrow
-		}
-		if fitsAll {
-			return flavor.Name, borrows, nil
-		}
-	}
-	return "", nil, &status
-}
-
-func flavorSelector(spec *corev1.PodSpec, allowedKeys sets.String) nodeaffinity.RequiredNodeAffinity {
-	// This function generally replicates the implementation of kube-scheduler's NodeAffintiy
-	// Filter plugin as of v1.24.
-	var specCopy corev1.PodSpec
-
-	// Remove affinity constraints with irrelevant keys.
-	if len(spec.NodeSelector) != 0 {
-		specCopy.NodeSelector = map[string]string{}
-		for k, v := range spec.NodeSelector {
-			if allowedKeys.Has(k) {
-				specCopy.NodeSelector[k] = v
-			}
-		}
-	}
-
-	affinity := spec.Affinity
-	if affinity != nil && affinity.NodeAffinity != nil && affinity.NodeAffinity.RequiredDuringSchedulingIgnoredDuringExecution != nil {
-		var termsCopy []corev1.NodeSelectorTerm
-		for _, t := range affinity.NodeAffinity.RequiredDuringSchedulingIgnoredDuringExecution.NodeSelectorTerms {
-			var expCopy []corev1.NodeSelectorRequirement
-			for _, e := range t.MatchExpressions {
-				if allowedKeys.Has(e.Key) {
-					expCopy = append(expCopy, e)
-				}
-			}
-			// If a term becomes empty, it means node affinity matches any flavor since those terms are ORed,
-			// and so matching gets reduced to spec.NodeSelector
-			if len(expCopy) == 0 {
-				termsCopy = nil
-				break
-			}
-			termsCopy = append(termsCopy, corev1.NodeSelectorTerm{MatchExpressions: expCopy})
-		}
-		if len(termsCopy) != 0 {
-			specCopy.Affinity = &corev1.Affinity{
-				NodeAffinity: &corev1.NodeAffinity{
-					RequiredDuringSchedulingIgnoredDuringExecution: &corev1.NodeSelector{
-						NodeSelectorTerms: termsCopy,
-					},
-				},
-			}
-		}
-	}
-	return nodeaffinity.GetRequiredNodeAffinity(&corev1.Pod{Spec: specCopy})
-}
-
-// fitsFlavorLimits returns whether a requested resource fits in a specific flavor's quota limits.
-// If it fits, also returns any borrowing required.
-func fitsFlavorLimits(rName corev1.ResourceName, val int64, cq *cache.ClusterQueue, flavor *cache.FlavorLimits) (int64, *admissionStatus) {
-	var status admissionStatus
-	used := cq.UsedResources[rName][flavor.Name]
-	if flavor.Max != nil && used+val > *flavor.Max {
-		status.AppendReason(fmt.Sprintf("borrowing limit for %s flavor %s exceeded", rName, flavor.Name))
-		return 0, &status
-	}
-	cohortUsed := used
-	cohortTotal := flavor.Min
-	if cq.Cohort != nil {
-		cohortUsed = cq.Cohort.UsedResources[rName][flavor.Name]
-		cohortTotal = cq.Cohort.RequestableResources[rName][flavor.Name]
-	}
-	borrow := used + val - flavor.Min
-	if borrow < 0 {
-		borrow = 0
-	}
-
-	lack := cohortUsed + val - cohortTotal
-	if lack > 0 {
-		lackQuantity := workload.ResourceQuantity(rName, lack)
-		if cq.Cohort == nil {
-			status.AppendReason(fmt.Sprintf("insufficient quota for %s flavor %s, %s more needed", rName, flavor.Name, &lackQuantity))
-		} else {
-			status.AppendReason(fmt.Sprintf("insufficient quota for %s flavor %s, %s more needed after borrowing", rName, flavor.Name, &lackQuantity))
-		}
-		// TODO(PostMVP): preemption could help if borrow == 0
-		return 0, &status
-	}
-	return borrow, nil
-}
-
 type entryOrdering []entry
 
 func (e entryOrdering) Len() int {
@@ -546,10 +289,10 @@ func (e entryOrdering) Less(i, j int) bool {
 	a := e[i]
 	b := e[j]
 	// 1. Request under min quota.
-	aMin := len(a.borrows) == 0
-	bMin := len(b.borrows) == 0
-	if aMin != bMin {
-		return aMin
+	aBorrows := a.assignment.Borrows()
+	bBorrows := b.assignment.Borrows()
+	if aBorrows != bBorrows {
+		return !aBorrows
 	}
 	// 2. FIFO.
 	return a.Obj.CreationTimestamp.Before(&b.Obj.CreationTimestamp)
@@ -570,14 +313,4 @@ func (s *Scheduler) requeueAndUpdate(log logr.Logger, ctx context.Context, e ent
 		}
 		s.recorder.Eventf(e.Obj, corev1.EventTypeNormal, "Pending", e.inadmissibleMsg)
 	}
-}
-
-func filterRequestedResources(req workload.Requests, allowList sets.String) workload.Requests {
-	filtered := make(workload.Requests)
-	for n, v := range req {
-		if allowList.Has(string(n)) {
-			filtered[n] = v
-		}
-	}
-	return filtered
 }
